[clean] separate definition of kernels
